+++ server/src/api/controllers/Admin/FetchShortUrlById.controller.ts
@@ -1,14 +1,14 @@
-import { ShortUrls } from "../models/shortURL.model";
+import { ShortUrls } from "../../models/shortURL.model";
 import { Response } from "express";
-import ErrorHandler from "../middleware/Handler/ErrorHandler";
-import HttpException from "../middleware/Exceptions/HttpException";
-import IMessage from "./interfaces/IMessage";
-import IResponse from "./interfaces/IResponse";
-import IShortUrl from "../models/interfaces/IShortUrl";
-import GetRequestProtocol from "../utils/GetRequestProtocol";
-import config from "../../config/env";
-import IRequest from "./interfaces/IRequest";
-import IReqParams from "./interfaces/IReqParams";
+import ErrorHandler from "../../middleware/Handler/ErrorHandler";
+import HttpException from "../../middleware/Exceptions/HttpException";
+import IMessage from "../interfaces/IMessage";
+import IResponse from "../interfaces/IResponse";
+import IShortUrl from "../../models/interfaces/IShortUrl";
+import GetRequestProtocol from "../../utils/GetRequestProtocol";
+import config from "../../../config/env";
+import IRequest from "../interfaces/IRequest";
+import IReqParams from "../interfaces/IReqParams";
 
 const FetchShortUrlsById = async (req: IRequest, res: Response) => {
   const msg: IMessage = { isSuccessful: false, message: "", info: null };